+++ internal/common/basic.go
@@ -5,9 +5,8 @@ import (
 	"fmt"
 	"time"
 
-	corev1 "k8s.io/api/core/v1"
-	capi "sigs.k8s.io/cluster-api/api/v1beta1"
-	capiexp "sigs.k8s.io/cluster-api/exp/api/v1beta1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	capi "sigs.k8s.io/cluster-api/api/core/v1beta2"
 	cr "sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/giantswarm/clustertest/v3/pkg/application"
@@ -202,7 +201,7 @@ func runBasic() {
 
 			mcClient := state.GetFramework().MC()
 			cluster := state.GetCluster()
-			Eventually(wait.IsClusterConditionSet(state.GetContext(), mcClient, cluster.Name, cluster.GetNamespace(), capi.ReadyCondition, corev1.ConditionTrue, "")).
+			Eventually(wait.IsClusterConditionSet(state.GetContext(), mcClient, cluster.Name, cluster.GetNamespace(), capi.ReadyCondition, metav1.ConditionTrue, "")).
 				WithTimeout(timeout).
 				WithPolling(wait.DefaultInterval).
 				Should(BeTrue(), failurehandler.LLMPrompt(state.GetFramework(), state.GetCluster(), "Investigate Cluster API Cluster CR not ready"))
@@ -274,7 +273,7 @@ func CheckWorkerNodesReady(ctx context.Context, wcClient *client.Client, values 
 // Running phase.
 func CheckMachinePoolsReadyAndRunning(ctx context.Context, mcClient *client.Client, clusterName string, clusterNamespace string) func() error {
 	return func() error {
-		machinePools := &capiexp.MachinePoolList{}
+		machinePools := &capi.MachinePoolList{}
 		machinePoolListOptions := []cr.ListOption{
 			cr.InNamespace(clusterNamespace),
 			cr.MatchingLabels{
@@ -295,14 +294,14 @@ func CheckMachinePoolsReadyAndRunning(ctx context.Context, mcClient *client.Clie
 		for _, mp := range machinePools.Items {
 			machinePool := mp
 			var machinePoolIsReady bool
-			machinePoolIsReady, err = wait.IsClusterApiObjectConditionSet(&mp, capi.ReadyCondition, corev1.ConditionTrue, "")
+			machinePoolIsReady, err = wait.IsClusterApiObjectConditionSet(&mp, capi.ReadyCondition, metav1.ConditionTrue, "")
 			if err != nil {
 				return err
 			}
 			allMachinePoolsAreReadyAndRunning = allMachinePoolsAreReadyAndRunning && machinePoolIsReady
 
-			currentMachinePoolPhase := capiexp.MachinePoolPhase(machinePool.Status.Phase)
-			machinePoolIsRunning := currentMachinePoolPhase == capiexp.MachinePoolPhaseRunning
+			currentMachinePoolPhase := capi.MachinePoolPhase(machinePool.Status.Phase)
+			machinePoolIsRunning := currentMachinePoolPhase == capi.MachinePoolPhaseRunning
 			allMachinePoolsAreReadyAndRunning = allMachinePoolsAreReadyAndRunning && machinePoolIsRunning
 			logger.Log(
 				"MachinePool '%s/%s' expected to be in Running phase, found MachinePool is in '%s' phase.",